Split operations into multiple modules.
